requirements.txt
streamlit==1.42.0
pillow==11.0.0
requests
